requirements.txt
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
certifi==2016.2.28
configparser==3.5.0
cssselect==0.9.1
decorator==4.0.9
entrypoints==0.2.2
functools32==3.2.3.post2
goose-extractor==1.0.25
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==5.1.5
jieba==0.38
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
lda==1.0.3
lxml==3.6.0
MarkupSafe==0.23
mistune==0.7.2
nbconvert==4.2.0
nbformat==4.0.1
nltk==3.2.1
notebook==4.2.0
numpy==1.11.0
pandas==0.18.1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.4
pyzmq==15.2.0
qtconsole==4.2.1
requests==2.10.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
tornado==4.3
traitlets==4.2.1
widgetsnbextension==1.2.3
wordcloud==1.2.1
